move common methods to mixin class
